Remove UTF module due to unsafe initialization requirements
The utf::initialize() function was calling svn_utf_initialize2() which
has strict thread safety requirements that cannot be safely met. The
documentation states "no other svn function may be in use in other
threads during the call of this function or when the pool is cleared
or destroyed", which is impossible to guarantee in multi-threaded tests.

Since the UTF conversion system cannot work without initialization and
initialization cannot be done safely, the entire UTF module has been
removed to prevent crashes.